pyproject.toml
[tool.poetry]
name = 'img2pptx'
version = '0.1.5'
description = ''
authors = ['MasterGowen <mastergowen@gmail.com>']
packages = [
    { include = 'img2pptx' }
]
repository = 'https://github.com/MasterGowen/img-to-pptx'
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Environment :: Console',
    'Intended Audience :: Information Technology',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.11',
    'Topic :: Multimedia :: Graphics :: Presentation',
    'Topic :: Software Development :: Build Tools',
]

[tool.poetry.dependencies]
python = '>3.7 <4.0'
python-pptx = '^0.6.21'
click = '^8.1.3'

[tool.poetry.scripts]
img2pptx = 'img2pptx.img2pptx:create_pptx'


[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
